styling and error handling

+++ frontend/src/components/newsfeed/index.js
@@ -63,7 +63,9 @@ const NewsFeed = () => {
 
   const handleDeletePost = (e, postId) => {
     e.preventDefault();
-    return dispatch(deletePost(postId));
+    if (window.confirm("are you sure?")){
+      return dispatch(deletePost(postId));
+    }
   }
 
   const handleEditPost = (e, post) => {
